# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0", 
    "title": "Ripio challenge",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "contact": {
      "name": "Agustin Mauras Velez",
      "email": "agustinmaurasvelez@gmail.com",
      "url": "https://www.linkedin.com/in/agustin-velez/"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "tags": [
    { "name": "Products" }
  ],
  "paths": {
    "/products": {
      "get" : {
        "tags": ["Products"],
        "description": "Get all products",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "id",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "status",
            "in": "query",
            "type": "integer",
            "required": false,
            "enum": [0, 1]
          },
          {
            "name": "owner",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "newOwner",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string",
            "required": false,
            "enum": ["id", "name", "status", "owner", "newOwner"]
          },
          {
            "name": "orderBy",
            "in": "query",
            "type": "string",
            "required": false,
            "default": "asc",
            "enum": ["asc", "desc"]
          },
          {
            "name": "refresh",
            "in": "query",
            "type": "boolean",
            "description": "Refresh cache memory",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/standard_response"
                }
              }
            }
          }
        },
        "produces": ["application/json"]
      },
      "post" : {
        "tags": ["Products"],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "description": "Product name",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "produces": ["text/plain"]
    },
    "/products/delegates/{addr}": { 
      "get" : {
        "tags": ["Products"],
        "description": "Get delegated products to an address",
        "parameters": [
          {
            "name": "addr",
            "required": "true",
            "in": "path",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "required": false
          },
          {
            "name": "id",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "status",
            "in": "query",
            "type": "integer",
            "required": false,
            "enum": [0, 1]
          },
          {
            "name": "owner",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "newOwner",
            "in": "query",
            "type": "string",
            "required": false
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string",
            "required": false,
            "enum": ["id", "name", "status", "owner", "newOwner"]
          },
          {
            "name": "orderBy",
            "in": "query",
            "type": "string",
            "required": false,
            "default": "asc",
            "enum": ["asc", "desc"]
          },
          {
            "name": "refresh",
            "in": "query",
            "type": "boolean",
            "description": "Refresh cache memory",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/standard_response"
                }
              }
            }
          }
        },
        "produces": ["application/json"]
      }
    },
    "/products/delegates": { 
      "post" : {
        "tags": ["Products"],
        "description": "Delegate a product to another account",
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "description": "Product id to delegate",
            "type": "integer"
          },
          {
            "name": "addr",
            "in": "body",
            "description": "Address of the delegate",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "produces": ["text/plain"]
    },
    "/products/delegates/accept": { 
      "post" : {
        "tags": ["Products"],
        "description": "Accept a product from another account",
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "description": "Product id to accept",
            "type": "integer"
          },
          {
            "name": "addr",
            "in": "body",
            "description": "Address of the acceptor",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "produces": ["text/plain"]
  },
  "components": {
    "schemas": {
      "standard_response": {
        "previous_page": {
          "type": "integer",
          "nullable": true,
          "description": "Previous page number"
        },
        "next_page": {
          "type": "integer",
          "nullable": true,
          "description": "Next page number"
        },
        "results": {
          "type": "array",
          "items": {
            "id": {
              "type": "integer"
            },
            "name": {
              "type": "string"
            },
            "status": {
              "type": "string",
              "description": ""
            },
            "owner": {
              "type": "string",
              "description": ""
            },
            "newOwner": {
              "type": "string",
              "description": ""
            }
          }
        }
      }
    }
  },
  "schemes": ["http"],
  "consumes": ["application/json"]
}